+++ packages/patternfly-4/react-core/src/components/Card/CardFooter.tsx
@@ -1,0 +1,25 @@
+import * as React from 'react';
+import styles from '@patternfly/patternfly/components/Card/card.css';
+import { css } from '@patternfly/react-styles';
+
+export interface CardFooterProps extends React.HTMLProps<HTMLDivElement> {
+  /** Content rendered inside the Card Footer */
+  children?: React.ReactNode;
+  /** Additional classes added to the Footer */
+  className?: string; 
+  /** Sets the base component to render. defaults to div */
+  component?: React.ReactNode;
+}
+
+export const CardFooter: React.FunctionComponent<CardFooterProps> = ({
+  children = null, 
+  className = '', 
+  component = 'div', 
+  ...props 
+}: CardFooterProps) => {
+  const Component = component as any;
+  return (
+  <Component className={css(styles.cardFooter, className)} {...props}>
+    {children}
+  </Component>
+)};