+++ packages/patternfly-4/react-core/src/components/List/List.tsx
@@ -1,0 +1,27 @@
+import * as React from 'react';
+import styles from '@patternfly/patternfly/components/List/list.css';
+import { css, getModifier } from '@patternfly/react-styles';
+
+export enum ListVariant {
+  inline = 'inline'
+}
+
+export interface ListProps extends React.HTMLProps<HTMLUListElement> {
+  /** Anything that can be rendered inside of the list */
+  children?: React.ReactNode;
+  /** Additional classes added to the list. */
+  className?: string;
+  /** Adds list variant styles */
+  variant?: 'inline';
+}
+
+export const List: React.FunctionComponent<ListProps> = ({
+  className = '',
+  children = null,
+  variant = null,
+  ...props
+}) => (
+  <ul {...props} className={css(styles.list, variant && getModifier(styles.modifiers, variant), className)}>
+    {children}
+  </ul>
+);