+++ packages/patternfly-4/react-integration/demo-app-ts/CHANGELOG.md
@@ -3,6 +3,59 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.1](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.10.0...demo-app-ts@1.10.1) (2019-05-21)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+# [1.10.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.9.0...demo-app-ts@1.10.0) (2019-05-20)
+
+
+### Features
+
+* **form:** Converted form component to typescript ([#1933](https://github.com/patternfly/patternfly-react/issues/1933)) ([110684d](https://github.com/patternfly/patternfly-react/commit/110684d))
+
+
+
+
+
+# [1.9.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.8.0...demo-app-ts@1.9.0) (2019-05-20)
+
+
+### Features
+
+* **Breadcrumb:** Convert Breadcrumb jsx file to tsx ([#1723](https://github.com/patternfly/patternfly-react/issues/1723)) ([9151262](https://github.com/patternfly/patternfly-react/commit/9151262))
+
+
+
+
+
+# [1.8.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.7.0...demo-app-ts@1.8.0) (2019-05-20)
+
+
+### Features
+
+* **button:** Converted button component to typescript ([#1903](https://github.com/patternfly/patternfly-react/issues/1903)) ([2b9d252](https://github.com/patternfly/patternfly-react/commit/2b9d252))
+
+
+
+
+
+# [1.7.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.6.0...demo-app-ts@1.7.0) (2019-05-20)
+
+
+### Features
+
+* **card:** Converted card component to typescript ([#1885](https://github.com/patternfly/patternfly-react/issues/1885)) ([e3e4170](https://github.com/patternfly/patternfly-react/commit/e3e4170))
+* **List:** Convert list to TypeScript ([#1946](https://github.com/patternfly/patternfly-react/issues/1946)) ([ecb2702](https://github.com/patternfly/patternfly-react/commit/ecb2702))
+
+
+
+
+
 # [1.6.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.5.1...demo-app-ts@1.6.0) (2019-05-17)